+++ src/fobi/contrib/plugins/form_elements/fields/email/forms.py
@@ -43,7 +43,7 @@ class EmailForm(forms.Form, BaseFormFieldPluginForm):
         required = False,
         widget = forms.widgets.Textarea(attrs={'class': theme.form_element_html_class})
         )
-    initial = forms.CharField(
+    initial = forms.EmailField(
         label = _("Initial"),
         required = False,
         widget = forms.widgets.TextInput(attrs={'class': theme.form_element_html_class})